+++ views/department/info.php
@@ -37,7 +37,7 @@ $this->params['breadcrumbs'] = [
 
         <?= $form->field($model, 'info')->widget(CKEditor::className()) ?>
         <?= Html::tag('span', '<i class="fa fa-fw fa-question"></i> Descreva o texto relacionado ao departamento.', ['id' => 'hbName', 'class' => 'help-block']) ?>
-        <?= Html::tag('span', '<i class="fa fa-fw fa-user"></i> Última alteração em ' . Yii::$app->getFormatter()->asDatetime($model->date_updated, 'long') . ' por ' . $model->getUserUpdated()->getName() . '.', ['class' => 'help-block']) ?>
+        <?= Html::tag('span', '<i class="fa fa-fw fa-user"></i> Última alteração em ' . Yii::$app->getFormatter()->asDatetime($model->date_updated, 'short') . ' por ' . $model->getUserUpdated()->getName() . '.', ['class' => 'help-block']) ?>
 
 
         <div class="form-group">
