+++ views/gallery/view.php
@@ -66,7 +66,7 @@ $this->params['breadcrumbs'] = [
             ],
             [
                 'attribute' => 'date_created',
-                'format' => ['datetime', 'long']
+                'format' => ['datetime', 'short']
             ],
             [
                 'attribute' => 'user_created',
@@ -74,7 +74,7 @@ $this->params['breadcrumbs'] = [
             ],
             [
                 'attribute' => 'date_updated',
-                'format' => ['datetime', 'long']
+                'format' => ['datetime', 'short']
             ],
             [
                 'attribute' => 'user_updated',
