+++ views/user/view.php
@@ -62,7 +62,7 @@ $this->params['breadcrumbs'] = [
             ],
             [
                 'attribute' => 'date_created',
-                'format' => ['datetime', 'long']
+                'format' => ['datetime', 'short']
             ],
             [
                 'attribute' => 'user_created',
@@ -70,7 +70,7 @@ $this->params['breadcrumbs'] = [
             ],
             [
                 'attribute' => 'date_updated',
-                'format' => ['datetime', 'long']
+                'format' => ['datetime', 'short']
             ],
             [
                 'attribute' => 'user_updated',
